refactor(SliderComponent): moved the slider component to shared folder and updated import paths
