+++ src/lib/components/shared/Slider/FullSlider/FullWithProgress.tsx
