+++ src/lib/components/shared/Slider/FullSlider/index.tsx
