+++ src/lib/components/shared/Slider/FullSlider/styles.ts
