+++ src/lib/components/shared/Slider/ProductSlider/index.tsx
