+++ src/lib/components/shared/Slider/index.tsx
