+++ src/lib/components/shared/index.ts
@@ -2,3 +2,4 @@ export * from './Button';
 export * from './Inputs';
 export * from './greeting';
 export * from './loaders';
+export * from './Slider';
